a large commit in re doing site media queries amongst other minor changes
This is the longest if commits were the media queries for the site had to be re done.

By using set breakpoints that cover a vast range of devices rather than specific ones.

Had a lot of difficulties across screen sizes but the bulk if not all the required outcomes have been met.

Minor changes have also been made by including a bold font weight to the nav links, changing colour on the text logo.